backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis & WebSocket
redis==5.0.1
websockets==12.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
qrcode[pil]==7.4.2
pillow==10.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# CORS - FastAPI has built-in CORS middleware
# No additional package needed

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Logging
structlog==23.2.0

# AI & Cloud Services
google-generativeai==0.3.1
cloudinary==1.36.0

# Background Tasks
celery==5.3.4
flower==2.0.1
